registration page css & class clean-up
Used bootstrap and a little extra css to standardize the look of the registration page so it matches the theme of the site. Remade the nested html classes so it's easier to work with in the future. Also modified the php a tiny bit to accept the cleaner html class names.